feat(fldi): add implementation handling
From now you can register a factory and you
will get the container as a parameter and
you can use different implementation for
your service.

Unfortunetly the dotmap is not immutatable
so later we should replace it with something
else, on the other hand it has own methods
like get and update, we need to use something
that has no own methods.